requirements.txt
# MCP Framework & Core Utilities
fastmcp>=3.2.3
pydantic>=2.12.5
uvicorn>=0.41.0
fastapi>=0.135.1

# Database & Persistence (For src/engines/memory/)
aiosqlite>=0.22.1
SQLAlchemy>=2.0.0

# Cognitive Analysis & NLP Metrics (For src/analysis/)
textstat>=0.7.13
numpy>=2.4.4
scikit-learn>=1.8.0
sentence-transformers>=5.4.0

# Network & Search (For src/tools/)
aiohttp>=3.13.3
httpx>=0.28.1
requests>=2.32.0

# Utilities
python-dotenv>=1.2.2
colorama>=0.4.6
tiktoken>=0.9.0

# Dashboard (Command Center)
streamlit>=1.56.0
pandas>=3.0.2
streamlit-autorefresh>=1.0.1